address peter's suggestion
