+++ projects/cm_mcu/I2CCommunication.h
@@ -16,6 +16,7 @@
 #include "FreeRTOSConfig.h"
 #include "stream_buffer.h"
 #include "common/smbus.h"
+#include "MonitorTask.h"
 
 // Command line interface
 int apollo_i2c_ctl_set_dev(uint8_t base);
@@ -23,5 +24,7 @@ int apollo_i2c_ctl_r(uint8_t address, uint8_t nbytes, uint8_t data[4]);
 int apollo_i2c_ctl_reg_r(uint8_t device, uint8_t address, uint8_t reg_address, uint8_t nbytes, uint8_t data[4]);
 int apollo_i2c_ctl_w(uint8_t address, uint8_t nbytes, int value);
 int apollo_i2c_ctl_reg_w(uint8_t address, uint8_t reg_address, uint8_t nbytes, int packed_data);
+int apollo_pmbus_rw(tSMBus *smbus, volatile tSMBusStatus *smbus_status, bool read,
+                    struct dev_i2c_addr_t *add, struct pm_command_t *cmd, uint8_t *value);
 
 #endif /* PROJECTS_CM_MCU_I2CCOMMUNICATION_H_ */
